pyproject.toml
[project]
name = "pcbre"
version = "0.1.0"
description = "Visual reverse-engineering for printed circuit boards."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Lukas Heintz" }]
keywords = ["pcb", "reverse-engineering", "electronics", "alignment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
dependencies = [
    "numpy>=1.20",
    "Pillow>=9.0",
]

[project.scripts]
pcbre = "pcbre.app:main"

[project.urls]
Homepage = "https://github.com/lukasloetkolben/pcbRE"
Source = "https://github.com/lukasloetkolben/pcbRE"
Issues = "https://github.com/lukasloetkolben/pcbRE/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pcbre"]
